mk_vendor_firmware: assert each function
* if any command fails, gtfo

+++ mk_vendor_firmware.sh
@@ -1,7 +1,15 @@
+assert(){
+if [ "$?" -ne "0" ]; then
+    echo -e "\nsub-process ${FUNCNAME[1]} failed. Bailing out!"
+    exit
+fi
+}
+
 unzip_ota(){
 rm -rf .workspace
 mkdir .workspace
 unzip $1 vendor.new.dat.br vendor.transfer.list vendor.patch.dat -d .workspace
+assert
 }
 
 if [ -z "$1" ]; then
